Move the files related to reviews, inquiries to "board" package
