+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/board/api/response/InquiriesResponse.kt
@@ -1,0 +1,7 @@
+package com.wafflestudio.toyproject.team4.core.board.api.response
+
+import com.wafflestudio.toyproject.team4.core.board.domain.Inquiry
+
+data class InquiriesResponse (
+    val inquiries: List<Inquiry>
+)
