+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/board/api/response/ReviewsResponse.kt
@@ -1,0 +1,7 @@
+package com.wafflestudio.toyproject.team4.core.board.api.response
+
+import com.wafflestudio.toyproject.team4.core.board.domain.Review
+
+data class ReviewsResponse(
+    val reviews: List<Review>
+)
