+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/board/database/ReviewImageEntity.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.team4.core.user.database
+package com.wafflestudio.toyproject.team4.core.board.database
 
 import javax.persistence.*
