+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/board/database/ReviewImageRepository.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.team4.core.user.database
+package com.wafflestudio.toyproject.team4.core.board.database
 
 import com.querydsl.jpa.impl.JPAQueryFactory
 import org.springframework.data.jpa.repository.JpaRepository
