+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/board/domain/Comment.kt
@@ -1,6 +1,6 @@
-package com.wafflestudio.toyproject.team4.core.user.domain
+package com.wafflestudio.toyproject.team4.core.board.domain
 
-import com.wafflestudio.toyproject.team4.core.user.database.CommentEntity
+import com.wafflestudio.toyproject.team4.core.board.database.CommentEntity
 import java.time.LocalDateTime
 
 data class Comment(
