--- src/main/kotlin/com/wafflestudio/toyproject/team4/core/item/api/response/InquiriesResponse.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.toyproject.team4.core.item.api.response
-
-import com.wafflestudio.toyproject.team4.core.item.domain.Inquiry
-
-data class InquiriesResponse (
-    val inquiries: List<Inquiry>
-)
