+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/item/service/ItemService.kt
@@ -1,15 +1,15 @@
 package com.wafflestudio.toyproject.team4.core.item.service
 import com.wafflestudio.toyproject.team4.common.CustomHttp404
-import com.wafflestudio.toyproject.team4.core.item.api.response.InquiriesResponse
+import com.wafflestudio.toyproject.team4.core.board.api.response.InquiriesResponse
 import com.wafflestudio.toyproject.team4.core.item.api.response.ItemRankingResponse
 import com.wafflestudio.toyproject.team4.core.item.api.response.ItemResponse
-import com.wafflestudio.toyproject.team4.core.item.database.InquiryRepository
+import com.wafflestudio.toyproject.team4.core.board.database.InquiryRepository
 import com.wafflestudio.toyproject.team4.core.item.database.ItemRepository
-import com.wafflestudio.toyproject.team4.core.item.domain.Inquiry
+import com.wafflestudio.toyproject.team4.core.board.domain.Inquiry
 import com.wafflestudio.toyproject.team4.core.item.domain.Item
-import com.wafflestudio.toyproject.team4.core.user.api.response.ReviewsResponse
-import com.wafflestudio.toyproject.team4.core.user.database.ReviewRepository
-import com.wafflestudio.toyproject.team4.core.user.domain.Review
+import com.wafflestudio.toyproject.team4.core.board.api.response.ReviewsResponse
+import com.wafflestudio.toyproject.team4.core.board.database.ReviewRepository
+import com.wafflestudio.toyproject.team4.core.board.domain.Review
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
