--- src/main/kotlin/com/wafflestudio/toyproject/team4/core/user/api/response/ReviewsResponse.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.toyproject.team4.core.user.api.response
-
-import com.wafflestudio.toyproject.team4.core.user.domain.Review
-
-data class ReviewsResponse(
-    val reviews: List<Review>
-)
