+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/user/database/PurchaseEntity.kt
@@ -1,6 +1,7 @@
 package com.wafflestudio.toyproject.team4.core.user.database
 
 import com.wafflestudio.toyproject.team4.core.item.database.ItemEntity
+import com.wafflestudio.toyproject.team4.core.board.database.ReviewEntity
 import org.springframework.data.annotation.CreatedDate
 import org.springframework.data.jpa.domain.support.AuditingEntityListener
 import java.time.LocalDateTime
